% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_identical_strings.R
\name{find_identical_strings}
\alias{find_identical_strings}
\title{A string vector}
\usage{
find_identical_strings(x, n = 2, specific = NULL)
}
\arguments{
\item{x}{a string vector}

\item{n}{how long}

\item{specific}{specific strings}
}
\description{
A string vector
}
