% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_loop_error.R
\name{ignore_loop_error}
\alias{ignore_loop_error}
\title{This is for ignore loop errors (caused by extreme scenario of resampling or something else).
You should run the code and know for sure the error is not a "hard error"}
\usage{
ignore_loop_error(
  rep,
  empty_name,
  some_operation_before = NULL,
  main_function,
  record_values,
  some_operation_after = NULL,
  output
)
}
\arguments{
\item{rep}{loop number}

\item{empty_name}{chunk of code, building empty object for saving results
the code should be written in "..
.. " With lines separate}

\item{some_operation_before}{chunk of code, do some manipualtion before the loop starts}

\item{main_function}{chunk of code, the main function to compute the result
the code should be written in "..
.. " With lines separate}

\item{record_values}{chunk of code, to save the result
the code should be written in "..
.. " With lines separate}

\item{some_operation_after}{chunk of code, do some manipualtion before the loop starts}

\item{output}{chunk of code, A list format the variables names you want to save as output "", should be something in the recorded values}
}
\value{
result
}
\description{
This is for ignore loop errors (caused by extreme scenario of resampling or something else).
You should run the code and know for sure the error is not a "hard error"
}
