% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_evaluation.R
\name{memory_evaluation}
\alias{memory_evaluation}
\title{evaluate memory useage of object and function Note that you do not do anything with the global environemnt.
running the code without running it}
\usage{
memory_evaluation(code, option = c("mem_change", "peakRAM"))
}
\arguments{
\item{code}{code chunk}

\item{option}{"mem_change","peakRAM"}
}
\description{
evaluate memory useage of object and function Note that you do not do anything with the global environemnt.
running the code without running it
}
