% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_loop.R
\name{parallel_loop}
\alias{parallel_loop}
\title{evaluate time and memory}
\usage{
parallel_loop(
  beforeloop_code,
  maincode,
  afterloop_code,
  parallel_option = c("none", "doParallel", "future")
)
}
\arguments{
\item{beforeloop_code}{code chunk}

\item{maincode}{code chunk}

\item{afterloop_code}{code chunk}

\item{parallel_option}{none future or do parallel}
}
\description{
evaluate time and memory
}
