% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_evaluation.R
\name{time_evaluation}
\alias{time_evaluation}
\title{evaluate time and memory}
\usage{
time_evaluation(
  code,
  how_many_times = 1,
  time_eval = c("system_time", "nanotime", "microbenchmark", "Rprof"),
  memory_eval = c("object_size")
)
}
\description{
evaluate time and memory
}
