% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_looptext_to_doParallel.R
\name{transform_loop_text_doParallel}
\alias{transform_loop_text_doParallel}
\title{Transform your loop text in for doparallel format}
\usage{
transform_loop_text_doParallel(x)
}
\arguments{
\item{x}{code chunk where you wrote a for loop}
}
\description{
Transform your loop text in for doparallel format
}
