% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_evaluation.R
\name{time_evaluation}
\alias{time_evaluation}
\title{evaluate time and memory}
\usage{
time_evaluation(code, how_many_times = 1, memory_eval = c("object_size"))
}
\arguments{
\item{code}{code chunk}

\item{how_many_times}{how many time you want to run the code chunk}
}
\description{
evaluate time and memory
}
